+++ src/Cache/FileCache.php
@@ -11,6 +11,7 @@ namespace HeimrichHannot\UtilsBundle\Cache;
 use Contao\File;
 use Contao\StringUtil;
 use HeimrichHannot\UtilsBundle\File\FileUtil;
+use Symfony\Component\DependencyInjection\ContainerInterface;
 use Symfony\Component\Filesystem\Filesystem;
 
 /**
@@ -36,7 +37,7 @@ class FileCache
      *
      * @var string
      */
-    protected $namespace;
+    protected $namespace = '';
 
     /**
      * The complete path to the current cache folder including namespace.
@@ -52,29 +53,40 @@ class FileCache
     /**
      * @var string
      */
-    private $webDir;
+    private $projectDir;
+    /**
+     * @var ContainerInterface
+     */
+    protected $container;
 
-    public function __construct(string $cacheFolder, string $webDir, FileUtil $fileUtil, string $namespace = '')
+    public function __construct(ContainerInterface $container)
     {
-        $this->cacheFolder = $cacheFolder;
-        $this->namespace = $namespace;
-        $this->fileUtil = $fileUtil;
-        $this->webDir = $webDir.'/..';
+        $this->cacheFolder = $container->getParameter('huh.utils.filecache.folder');
+        $this->fileUtil    = $container->get('huh.utils.file');
+        $this->projectDir  = $container->getParameter('kernel.project_dir');
         $this->generatePath();
+        $this->container = $container;
     }
 
     /**
-     * Checks if a cached file already exist.
+     * Checks if a cached file already exist in cache. Namespace is taken into account.
      *
-     * @param string $identifier    The identifier
-     * @param string $fileExtension
+     *
+     * @param string $identifier The identifier
+     * @param string $fileExtension If not set, a file with no file extension is searched
      *
      * @return bool
+     * @throws \Exception
      */
-    public function exist(string $identifier, string $fileExtension)
+    public function exist(string $identifier, string $fileExtension = '')
     {
         $fileName = $this->getCacheFileName($identifier);
-        $file = new File($this->cacheFolderWithNamespace.'/'.$fileName.'.'.$fileExtension);
+        $cachePath  = $this->cacheFolderWithNamespace . '/' . $fileName;
+        if(!empty($fileExtension))
+        {
+            $cachePath .= '.'.$fileExtension;
+        }
+        $file     = new File($cachePath);
 
         if ($file->exists()) {
             return true;
@@ -84,18 +96,26 @@ class FileCache
     }
 
     /**
-     * Get the file path for the given identifier.
+     * Get the file path for the given identifier. Namespace is taken into account.
      *
-     * @param string   $identifier
-     * @param string   $fileExtension
-     * @param callable $saveCallback  A callback handles the file save functionality. Get filepath, filename and the identifier as parameter. Expects a boolean return value.
+     * @param string $identifier
+     * @param string $fileExtension
+     * @param callable $saveCallback A callback handles the file save functionality. Get filepath, filename and the identifier as parameter. Expects a boolean return value.
      *
-     * @return bool|string
+     * @return bool|string Returns the path of the cached file or false, if cached file could not be found.
+     * @throws \Exception
      */
-    public function get(string $identifier, string $fileExtension, callable $saveCallback = null)
+    public function get(string $identifier, string $fileExtension = '', callable $saveCallback = null)
     {
-        $fileName = $this->getCacheFileName($identifier).'.'.$fileExtension;
-        $file = new File($this->cacheFolderWithNamespace.'/'.$fileName);
+        $fileName = $this->getCacheFileName($identifier);
+        $cachePath  = $this->cacheFolderWithNamespace . '/' . $fileName;
+        if(!empty($fileExtension))
+        {
+            $fileName .= '.'.$fileExtension;
+            $cachePath .= '.'.$fileExtension;
+
+        }
+        $file     = new File($cachePath);
 
         if (!$file->exists()) {
             if (null !== $saveCallback) {
@@ -173,7 +193,7 @@ class FileCache
      */
     public function getAbsoluteCachePath()
     {
-        return $this->webDir.'/'.$this->cacheFolderWithNamespace;
+        return $this->projectDir.'/'.$this->cacheFolderWithNamespace;
     }
 
     /**
@@ -231,13 +251,14 @@ class FileCache
     {
         $filesystem = new Filesystem();
         $path = $this->cacheFolder;
+        $path = trim($path, "/");
 
         if (!empty($this->namespace)) {
             $path .= '/'.$this->namespace;
         }
 
-        if (!$filesystem->exists($this->webDir.'/'.$path)) {
-            $filesystem->mkdir($this->webDir.'/'.$path);
+        if (!$filesystem->exists($this->projectDir.'/'.$path)) {
+            $filesystem->mkdir($this->projectDir.'/'.$path);
         }
         $this->cacheFolderWithNamespace = $path;
     }